streamlit/requirements.txt
streamlit
numpy
Pillow
scikit-learn
opencv-python-headless